% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixar_paginas_ponte.R
\name{baixar_paginas_ponte}
\alias{baixar_paginas_ponte}
\title{Baixar páginas de notícias da Ponte Jornalismo}
\usage{
baixar_paginas_ponte(diretorio = ".", paginas = 10)
}
\arguments{
\item{diretorio}{diretorio onde serão baixadas as notícias}

\item{paginas}{Número de páginas}
}
\value{
html
}
\description{
Baixar páginas de notícias da Ponte Jornalismo
}
