% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ler_paginas_ponte.R
\name{ler_paginas_ponte}
\alias{ler_paginas_ponte}
\title{Ler notícias baixadas com baixar_paginas_ponte}
\usage{
ler_paginas_ponte(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor com local dos arquivos}

\item{diretorio}{Se não informar os arquivos,
informar o diretório}
}
\value{
tibble
}
\description{
Ler notícias baixadas com baixar_paginas_ponte
}
